Move significant functions to a separate module
